Make tests build snap localy
This commit first fixes how the application under tests is
executed. Due to the spread design it is necessary to use absolute
paths, so starting with /snap/bin, rather than just the application
name, such as pulseaudio.pactl. Second, this commit makes the snap
to be build locally and installed.